+++ apps/hyperdrive-trading/src/ui/appconfig/useAppConfigForConnectedChain.ts
@@ -3,7 +3,7 @@ import {
   isMainnetChain,
   mainnetAppConfig,
   testnetAppConfig,
-} from "@hyperdrive/appconfig";
+} from "@delvtech/hyperdrive-appconfig";
 import { useChainId } from "wagmi";
 
 export function useAppConfigForConnectedChain(): AppConfig {