+++ apps/hyperdrive-trading/src/ui/chainlog/Chainlog.tsx
@@ -1,5 +1,5 @@
+import { appConfig, makeAddressUrl } from "@delvtech/hyperdrive-appconfig";
 import { ArrowTopRightOnSquareIcon } from "@heroicons/react/24/outline";
-import { appConfig, makeAddressUrl } from "@hyperdrive/appconfig";
 import { useNavigate, useSearch } from "@tanstack/react-router";
 import { ReactElement } from "react";
 import { Tabs } from "src/ui/base/components/Tabs/Tabs";