+++ apps/hyperdrive-trading/src/ui/hyperdrive/hooks/useLpApy.ts
@@ -1,4 +1,7 @@
-import { appConfig, findHyperdriveConfig } from "@hyperdrive/appconfig";
+import {
+  appConfig,
+  findHyperdriveConfig,
+} from "@delvtech/hyperdrive-appconfig";
 import { useQuery } from "@tanstack/react-query";
 import { makeQueryKey } from "src/base/makeQueryKey";
 import { getLpApy, LpApyResult } from "src/hyperdrive/getLpApy";