+++ apps/hyperdrive-trading/src/ui/markets/PoolDetails.tsx
@@ -1,5 +1,5 @@
+import { appConfig, HyperdriveConfig } from "@delvtech/hyperdrive-appconfig";
 import { ArrowLeftIcon } from "@heroicons/react/16/solid";
-import { appConfig, HyperdriveConfig } from "@hyperdrive/appconfig";
 import { Link, useSearch } from "@tanstack/react-router";
 import classNames from "classnames";
 import { ReactElement } from "react";